Adding WebUtil class for web & dom related startup tasks
